+++ src/App.svelte
@@ -16,6 +16,7 @@
   import VocabularyRichness from './components/VocabularyRichness.svelte';
   import SentenceRhythmInspector from './components/SentenceRhythmInspector.svelte';
   import TextAnonymizer from './components/TextAnonymizer.svelte';
+  import TextSummarizer from './components/TextSummarizer.svelte';
   import TextDiffViewer from './components/TextDiffViewer.svelte';
   import ReadabilityScore from './components/ReadabilityScore.svelte';
   import KeywordDensity from './components/KeywordDensity.svelte';
@@ -49,6 +50,8 @@
   <SentenceRhythmInspector />
 
   <TextAnonymizer />
+
+  <TextSummarizer />
 
   <TextDiffViewer />